+++ src/main/java/uz/uat/app/intercom/model/entity/channels/Channel.java
@@ -1,0 +1,23 @@
+package uz.uat.app.intercom.model.entity.channels;
+
+import java.time.LocalDate;
+
+import org.springframework.data.annotation.Id;
+
+import com.arangodb.springframework.annotation.ArangoId;
+import com.arangodb.springframework.annotation.Document;
+
+import lombok.Data;
+
+@Data
+@Document(collection = "chanels")
+public class Channel {
+    @Id
+    private String id;
+    @ArangoId
+    private String arangoId;
+    private String image;
+    private String name;
+    private LocalDate startDate;
+    private String status;
+}
