+++ src/main/java/uz/uat/app/intercom/model/entity/channels/ChannelTypes.java
@@ -1,8 +1,8 @@
-package uz.uat.app.intercom.model.entity.chats;
+package uz.uat.app.intercom.model.entity.channels;
 
 import com.google.common.collect.ImmutableList;
 
-public class ChanelTypes {
+public class ChannelTypes {
         public static final String USER = "USER";
         public static final String INFO = "INFO";
         public static final String ROBOT = "ROBOT";
