+++ src/main/java/uz/uat/app/intercom/model/entity/messages/Draft.java
@@ -1,6 +1,4 @@
-package uz.uat.app.intercom.model.entity.chats;
-
-import java.time.LocalDate;
+package uz.uat.app.intercom.model.entity.messages;
 
 import org.springframework.data.annotation.Id;
 
@@ -10,14 +8,10 @@ import com.arangodb.springframework.annotation.Document;
 import lombok.Data;
 
 @Data
-@Document(collection = "chanels")
-public class Channel {
+@Document(collection = "drafts")
+public class Draft extends Message {
     @Id
     private String id;
     @ArangoId
     private String arangoId;
-    private String image;
-    private String name;
-    private LocalDate startDate;
-    private String status;
 }
